[sshd] Authentication + Authorization
Changed the setup to 2 different servers. One is
a secure server that requires an Authority/AuthorizationAdmin
and Authenticator to be present. The other can only
be registered on localhost and has no authentication.

Both servers _must_ be configured with a factory instance.

The imports for the secure server are optional so
you can run into problems.

The insecure and secure servers have different configurations.