[jna] Proper time of GC
JNA creates native memory (not sure
actually why it needs to be native) for
some structures. This memory is
freed when the owner is garbage
collected. 

This makes it necessary to run the
GC before the library is closed.

Added some code to be able to see
how memory is allocated when the native
code creates the struct or when Java
creates the struct.

Noticed that if the Java struct created in native code is not
specifically set to not synchronize then it reads memory of a free pointer.