[mqtt] Redesigned API
The API was using the whiteboard but this conflicted
with the primary use case that this design addressed.
In this use case, it is necessary for the code
to create random topics.

So now it is a traditional broker model which
looks pretty close to the PAHO client itself.

Also added a test project to use LaunchPad with
a real test.