[logger] Additions and fixes to LoggerAdmin commands
Tracking latest errors and warnings – If the log is heavily abused with debug messages, the errors and warnings tend to be missed. This changes keeps track of errors and warnings  separately. This makes it easy to see the last errors.

Explicitly uses null for accessing the root context. The name ROOT did not work although this is according to spec as far as I can see.