[connection] #22 MQTT Client Cannot Reconnect
Created a test case to interrupt the connection
as described. Was not able to get into the reported
state.

What is needed is a full FINE log of the client, this
will likely make it possible to to create a test
case that reflects this.

Attempt to replicate the effect with the reporter
failed.

To be able to add more tweaking, new configuration parameters
were added.


Link https://github.com/aQute-os/biz.aQute.osgi.util/issues/22