[tests] Tests were failing
For some reason tests were not run on the 
CI server. So when I found this out, it turned
out there were quite a few failing tests.

One of the reasons was the disastrous change
in aQute.libg in 6.1. It renamed a package
and then made it available in a util jar.

For now, reverted to 5.3.0, fixed the package
in likely bnd 6.4 ... (maybe can push it to 
6.3)