[fixup] Do an early exit before the fixup when level not enabled
The fixup of arguments is there to allow proper printing of arrays. It could take significant time if the arrays were relatively large. This happened before
we checked the log level.

This now will not happen when the level is not active. However, before the
log service is available, ALL levels are active.