[exchange] Exchange class could underflow
Stupid braindamaged error ... A counter was
maintained to keep the # of next bytes for
a codepoint. This counter was decremented
and then tested to be > 0 ....

Works fine if now and then there is a character
encoded with > 1 byte ... unfortunately if there
are 2Gb of characters without even one being encoded
in > 1 byte then the counter underflowed and
the code then assumed the normal character needed 
multiple byte encoding ...

The solution was simple, just test the counter 
before decrementing ...

Stupid stupid stupid.